Add typescript. Move server files to subfolder. Use env variable to parse version
