+++ index.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2018 Orange
+ *
+ * This software is the confidential and proprietary information of Orange.
+ * You shall not disclose such confidential information and shall
+ * use it only in accordance with the terms of the license agreement you
+ * entered into with Orange.
+ */
+
+import {Server} from "./server/server";
+
+const server = new Server();
+server.start();
