+++ package.json
@@ -19,15 +19,14 @@
     }
   ],
   "engines": {
-    "node": ">= 0.10.0",
-    "npm": ">= 1.0.0"
+    "node": ">=12.16"
   },
   "dependencies": {
     "body-parser": "^1.19.0",
     "boosted": "4.0.0-alpha.4",
     "browserify": "14.1.0",
     "clipboard": "1.6.1",
-    "debug": "2.6.9",
+    "debug": "4.3.2",
     "express": "^4.17.1",
     "jquery": "3.5.0",
     "jquery-file-download": "1.4.6",
@@ -38,13 +37,17 @@
     "socket.io-client": "2.3.0",
     "socket.io-stream": "0.9.1",
     "tether": "1.1.1",
-    "uglify-js": "2.8.14"
+    "uglify-js": "3.14.1",
+    "typescript": "^3.7.5",
+    "ts-node": "^5.0.1"
   },
   "scripts": {
-    "start": "",
+    "start": "ts-node index.ts",
     "test": "mocha "
   },
   "devDependencies": {
+    "@types/debug": "^4.1.7",
+    "@types/node": "^16.7.1",
     "conventional-changelog-cli": "~1.3.1",
     "mocha": "^7.1.2",
     "mockery": "~2.0.0",
@@ -53,5 +56,5 @@
   },
   "bundleDependencies": [],
   "private": true,
-  "main": "server.js"
+  "main": "server/server.ts"
 }
