+++ server/public/js/HackTimerWorker.js
