+++ server/views/receive.pug
